VPA: Allow admission-controller to update in-place
With the InPlacePodVerticalScaling feature, we can now update the
resource requests and limits of pods in-place rather than having to
evict the pod to rewrite it.

We do have to make sure though ( because apps could react badly to an
update or require container restarts ) that we limit the amount of
disruption we can introduce at once, so we limit our updates only to the
ones that the updater has okayed.

(And then, over in the updater we're going to meter them so they don't
get sent to the admission-controller all at once)

This commit:
- allows the admission-controller to monitor update operations
- adds the new UpdateOrRecreate update mode to the list of possible
  update modes
- makes sure the admission-controller only patches pod update requests
  that were triggered by the updater (by using a special annotation)
- makes sure the admission-controller removes the annotation upon
  patching to signify that it is done