VPA: (squash) Fix updater test for in-place
So in-place VPA checks happen before eviction checks do, and we fall
back to eviction if in-place VPA isn't possible.

In-place VPA happens immediatly (rather than the delay with eviction) so
we seem to be able to catch early "0" recommendations from the VPA when
it hasn't recommended anything yet unless we wait for the VPA's
recommendation to be provided.

The problem is, the updater test was not set up to accommidate that, it
expects eviction immediately which would seem to be wrong and probably
not what we want (it's possible something else before in-place made
that "not wrong" but for now I'm operating under the assumption that the
"eviction delay" was saving us from having any trouble with this in the
real world).

So until I get to the bottom of how we should be behaving when the VPA
hasn't made a recommendation yet, I'm going to adjust the test suite so
it allows for in-place's "wait until we have a recommendation" logic.

And I need to come back here and have proper tests for in-place with
simulating the InPlacePodVerticalScaling behavior being enabled.